Updates to relative journal page links
For Journal Pages embedded in Journal when the link path is specified relative to the parent, the normal link creation function didn't work. Had to modify the target link to get the uuid by searching for the embedded journal pages uuid. Probably there's a smarter way to get at this, and it might break if other embedded object types show up.